Don't call virtual functions from the constructor
In order to increase the modularity of the kernel, I have turned the block reward and proof of work functions into pure virtual functions which have to be difined by the user of the kernel. However, since these two functions were used by the constructor (which is not allowed by C++), I had to separate blockchain loading into its own function which is called after the constructor.