Ensure the input id is correct when verifying a transaction
This is to prevent a potential attack where a user submits a transaction using an input with a different value or owner than it originally had. This could allow an attack during a chain reorg where the input as specified in the block is different to how it was issued in a previous block. This could mean that as the block is reversed, the input is added back to the UTXO set with a different value than its original value when it was spent the first time.