Implement importprivkey RPC call
This RPC allows you to import an external private key into a CK wallet. Doesn't yet rescan to take account of new balances or recognise if the key is already in the database. There is some bug in the Crypto class which prevents it from deriving the pubKey from the privKey.