Fix mempool uncaught exception (#16)
* Fix bug where spending from the wallet soon after a reorg could cause an uncaught NotFoundException

* Fix bug where mempool would get corrupted after a reorg

Each transaction input has an associated output ID that it is spending. These were not being included in the mempool conflict checking and thus it was possible for the mempool to get corrupted if a reorg takes place or if transactions spending the same output IDs were submitted. This caused an uncaught InvalidElementException in the validating block generation code.

* Build fixes for armhf