Fix deadlock caused during Lua script verification
Not entirely sure what's causing this to happen. I though Lua code was executed in the same thread the execution was started from but perhaps not. Previously the code was blocking on a lock it was already has. As a workaround just remove the lock, it's kind of superfluous anyway since the Storage::Transaction already has a lock on the blockchain class' state.