Fix bug where every second block poll's blocks would not get submitted
If there had previously been a blockProcessor thread running, the cleanup code would then see the currentHeight equal bestHeight because the network thread had downloaded the latest blocks. Remove that check becuase it's not needed and currentHeight does not refer to the current blockchain height.

Additionally add some more logging to the block download thread that was useful in debugging this bug.