Implement a read-only storage transaction and cache peerInfo for RPC (#45)
Previously lock contention was very high while the blockchain was syncing. This meant many RPC calls would timeout. Instead make use of leveldb snapshots and peerInfo caching internally to the blockchain and network to prevent waiting on several locks. This is at the cost of slightly stale data from RPC.