Improve transaction commit speed by changing the publicKey index method (#48)
Previously to keep track of the map of publicKeys to unspent and spent outputs we used a JSON array with one key per public key. If a given public key was used many times, this would lead to a very large JSON object being retrieved from and saved to the database for every output using that public key. This lead to cache misses and large wait times for database operations.

Instead, play to leveldb's strengths by using one key per output and retreiving the list of outputs for a public key using an iterator. This means the speed to update the public key/output map is now almost constant rather than linear with the number of outputs associated with a public key.

The PR also adds a unit test for basic getUnspentOutputs functionality.