Signature aggregation in transactions (#58)
This PR implements MuSig schnorr signature aggregation for transactions. This allows for multiple schnorr inputs to be signed with a single signature. To sign a group of schnorr outputs with an aggregate signature, leave the data field of each schnorr input to be signed without a signature field except for one input where you instead include an aggregateSignature.

The aggregateSignature takes the following form:

"aggregateSignature": {
    "signs": [0, 3, 5],
    "signature": <base64 encoded sig>
}

The signs field enumerates the outputs that this aggregate signature covers. They are indexed by ordering all the output IDs spent in the transaction from smallest to largest. Each schnorr output may be covered by at most one aggregate signature.

The signed message payload is the ordered output IDs concatenated together with the output set merkle root.