Lock the connection while we are initialising it so that it doesn't get prematurely used in another thread (#69)
If a peer with a very small ping connected to ckd then the peer could become disconnected in another thread and the connection deleted before its initial info was committed to the database. That resulted in a segmentation fault in some cases where the connection was deleted in another thread and then subsequently used in the initialisation thread. This PR locks the connection right after its creation and only releases it once the peer is fully set up. The PR also removes an unnecessary sleep call.