Pay to merkleroot (#62)
This implements a "pay to merkleroot" transaction type.

When creating the output, we only specify the merkleRoot value to be the merkleroot of all possible pubkeys/scripts allowed spending the output.

When spending the output, we include the following in the spendData:
spendType - this is "pubkey" or "script" 
pubKeyOrScript - containing the original pubKey or script used to spend
merkleProof - contains the proof that the pubkey or script is part of the tree. Contains all the leaves required to calculate a matching merkle root. This is the .toJson() of the MerkleProof class
signature - when using the spendType="pubkey" this field contains the signature of that key signing the transaction.