Don't send invalid IP addresses in info messages
Sometimes nodes get connected to themselves, or are unable to resolve IP addresses. In that case and IP address such as 127.0.0.1 or 0.0.0.0 might be sent, which will cause the remote peer to disconnect. Add a check for such IP addresses before adding them to the peers list to send with an info message. Also improved the log message for disconnecting a peer behaving in this way to display the offending malformed IP address.