Catch script resource exceptions to prevent crashes (#86)
If a Lua script was being validated and exceeded the allowed memory the allocator would throw an exception causing ckd to crash. This PR catches the exception so the script validation function returns false which is the expected behaviour. Also add some unit tests to ensure the correct behaviour in the event of script execution errors.