Catch InvalidElementException when generating a new mining block (#87)
Because the mempool is not saved in the database, it can become inconsistent with the current state of the UTXO set. In this case, calculating the transaction fees can cause an exception to be thrown as some of the transaction in our unconfirmed set become invalid and their inputs are removed from the database. As a workaround, simply return a block with no transactions in it.